+++ resources/views/components/attendance/member-group-tree-record.blade.php
@@ -8,7 +8,7 @@
     <div x-on:click="show= !show" :class="show ? 'bg-indigo-100':''"
          class="cursor-pointer transition flex justify-between sm:justify-start space-x-4 gap-2 py-2 px-5 items-center hover:bg-indigo-200 rounded">
         <i class="fas " :class="show ? 'fa-minus':'fa-plus'"></i>
-        <h2 class="text-xl">{{__($memberGroup->title)}}</h2>
+        <h2 class="md:text-xl">{{__($memberGroup->title)}}</h2>
     </div>
     <!-- Content -->
     <div x-show="show" class="px-5 pt-0 overflow-hidden" x-transition>