+++ resources/views/livewire/attendance/single-attendance.blade.php
@@ -6,32 +6,40 @@
     $attendance = $event->attendances()->where("member_id", $member->id)->first();
 @endphp
 <div class="flex flex-col justify-between sm:flex-row py-2">
-    <div class="p-2">{{$member->getFullName()}}</div>
-    <div class="flex justify-center gap-2">
-        <div wire:click="recordPoll({{$attendance?->poll_status === "in" ? 'null' : "'in'"}})" title="{{__("In")}}"
+    <div class="p-2 max-md:text-sm">{{$member->getFullName()}}</div>
+    <div class="relative flex justify-center gap-2">
+        <div wire:click="recordPoll({{$attendance?->poll_status === "in" ? 'null' : "'in'"}})"
+             title="{{__("In")}}"
              class="rounded-full w-10 h-10 flex justify-center items-center cursor-pointer
                         {{$attendance?->poll_status === "in" ? 'text-white bg-green-700' : 'text-green-700'}}">
             <i class="fa-solid fa-check"></i>
         </div>
 
-        <div wire:click="recordPoll({{$attendance?->poll_status === "unsure" ? 'null' : "'unsure'"}})" title="{{__("Unsure")}}"
-            class="rounded-full w-10 h-10 flex justify-center items-center cursor-pointer
+        <div wire:click="recordPoll({{$attendance?->poll_status === "unsure" ? 'null' : "'unsure'"}})"
+             title="{{__("Unsure")}}"
+             class="rounded-full w-10 h-10 flex justify-center items-center cursor-pointer
                         {{$attendance?->poll_status === "unsure" ? 'text-white bg-yellow-700' : 'text-yellow-700'}}">
             <i class="fa-solid fa-question"></i>
         </div>
 
-        <div wire:click="recordPoll({{$attendance?->poll_status === "out" ? 'null' : "'out'"}})" title="{{__("Out")}}"
-            class="rounded-full w-10 h-10 flex justify-center items-center cursor-pointer
+        <div wire:click="recordPoll({{$attendance?->poll_status === "out" ? 'null' : "'out'"}})"
+             title="{{__("Out")}}"
+             class="rounded-full w-10 h-10 flex justify-center items-center cursor-pointer
                         {{$attendance?->poll_status === "out" ? 'text-white bg-red-700' : 'text-red-700'}}">
             <i class="fa-solid fa-xmark"></i>
         </div>
 
-
         <div class="border-r border-gray-500"></div>
-        <div wire:click="recordAttend({{$attendance?->attended ? 'null' : 'true'}})" title="{{__("Attended")}}"
-            class="rounded-lg w-10 h-10 flex justify-center items-center cursor-pointer
+        <div wire:click="recordAttend({{$attendance?->attended ? 'null' : 'true'}})"
+             title="{{__("Attended")}}"
+             class="rounded-lg w-10 h-10 flex justify-center items-center cursor-pointer
                         {{$attendance?->attended ? 'text-white bg-green-700' : 'text-green-700'}}">
             <i class="fa-solid fa-clipboard-check"></i>
+        </div>
+
+        <div wire:loading.flex
+             class="flex flex-row justify-center items-center absolute left-0 right-0 top-0 bottom-0 select-none z-50 bg-gray-300 opacity-50">
+            <i class="fa-solid fa-spinner fa-spin-pulse text-lg"></i>
         </div>
     </div>
 </div>