+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 ### Changed
 - use standard notification message for click on calendar link on dashboard
 - store member filter selection in database instead of browser
+- if members are in multiple groups the other components get now update if one changes in attendance record
 
 ## v0.6.4 - 2024-02-11
 ### Added